Fix PATH issues
The code for setting the PATH variable had potential errors depending on when
the functions were called, to fix this we use the funky bash syntax to ensure
setting the path is correct and not a security issue.

There was also a bug where we would add a new instance of the TeXLive path
for each invoked subshell, this is now fixed.